feat: Test the actual receptive field
The receptive field size is calculated using a closed formula. Added
a test to check the partial derivatives of the input w.r.t. the last
output. The last `receptive_field` steps are expected to have non-zero
derivatives, and all the remaining leading steps are expected to have
zero derivatives. Tested.