feat: More flexible Tracks dataset
Added overlapping framing. This will probably be needed for the improved
batching ticket. It also lets us train on larger datasets with
overlapped frames, where the overlap is as large as the receptive field.
I think that this will already improve performance.

Also added a cache key for the caching, so that we can change audio
parameters without remembering to manually wipe the cache. This sort of
cache corruption may be hard to detect, and could completely mess with
experimentation.